configs/*demo*: remove Dropbear as OpenSSH is also selected
Enabling both Dropbear and OpenSSH in the same defconfig does not make
a lot of sense, as they both provide a SSH server and client. Only one
of them will act as a SSH server, as only one of them can listen to
TCP port 22.

Since OpenSSH is more feature-complete and those demo defconfigs
already have lots of other packages, keeping OpenSSH seems the best
choice.

This commit solves a comment made as part of
https://github.com/linux4sam/buildroot-external-microchip/issues/1.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>