board: sama5: defconfig: Change FORCE_MAX_ZONEORDER to default value
The minimum size of the CMA can't be lower than the maximum size
available for a contiguous memory block, due to alignment constraints.
FORCE_MAX_ZONEORDER is used to select the largest power of two that the
kernel keeps in the memory allocator. With its value of 15, the minimum
value becomes PAGE_SIZE * 2^14 = 64MB. This buffer is to big for
sama5d27-som1-ek board and the CMA allocation fails. We should use the
default value instead, which is 11.

Signed-off-by: Codrin Ciubotariu <codrin.ciubotariu@microchip.com>