configs: sama5d2_xplained_greenkey: Replace systemd with eudev
systemd was used in the previous setup because of a problem with
the AWS IoT Device Tester software. The current version of the
Device Tester software does not have this same issue, and systemd
is NOT required.

The serial console was broken when using systemd. Replacing systemd
with eudev solves the problem.

Signed-off-by: Tudor Ambarus <tudor.ambarus@microchip.com>