board: sama5d{3,4}-xplained: u-boot-env: remove atmel.pm_modes
SAMA5D4 Xplained and SAMA5D3 Xplained available PM modes are standby
and ULP0. Remove atmel.pm_modes bootcmd variable so that the user
to not be tempted to change it. If no atmel.pm_modes variable is
present in bootcmd the default PM modes (standby, ULP0) will be used.

Signed-off-by: Claudiu Beznea <claudiu.beznea@microchip.com>
[eugen.hristev@microchip.com: amended shortlog]
Signed-off-by: Eugen Hristev <eugen.hristev@microchip.com>