Revert "configs: sama5d2: Enable OPENJDK for AWS GG"
This reverts commit 9157dc5f5d7752171401394e5812994fe45a0890.

Remove the java support that was intended for AWS GG. The reasons are:
1/ we couldn't make the Stream Manager work neither on yocto (with java8),
nor on buildroot (java 14).
2/ difference between the java versions on yocto (java 8) and buildroot
(java11 or java14)
3/ java11 or java14 depend on x11, the image size increased with
~110 MBytes, the rootfs becomes too tight for our 512 MByte NAND on
sama5d2_ptc.
4/ there was no requirement to support lamdas with java, I experienced this
just for fun :).

One will notice that the graphics images introduce
BR2_PACKAGE_XUTIL_UTIL_MACROS. This config was selected by
BR2_PACKAGE_XORG7. With the removal of XORG7, we have to introduce back
BR2_PACKAGE_XUTIL_UTIL_MACROS for the graphics images.

Signed-off-by: Tudor Ambarus <tudor.ambarus@microchip.com>