board/sama7g5ek: add fs-overlay for camera booth
These collection of scripts and configuration enables the sama7g5ek to act
as a Photo Camera.
NGINX server is started at boot.
PB_USER button is used for photo capture. Push the button and you have a
new photo. Old photos are not overwritten. Photo name includes the date it was
taken.
All photos are stored in /usr/html/captures
All photos are available by web, if you configure Ethernet and connect
with a browser to the board IP address, then you will have a 'captures'
directory with all taken photos.
A script that waits for push button and takes photos is automatically
started at boot time.

Signed-off-by: Eugen Hristev <eugen.hristev@microchip.com>