sama5d2: fs-overlay: AWS GG: Replicate greengrass-init from meta-aws
Without much time for replicating meta-aws, we polish what we have for now.

Add the script just for the sama5d2 boards and not for all sama5 boards,
as Eugen suggested.

Move the start script in /etc/init.d/ and rename it to "greengrass". One
should pass one of the {start, stop, restart} arguments.

Signed-off-by: Tudor Ambarus <tudor.ambarus@microchip.com>