package/wilcbtapps: rework Makefile to avoid possible errors
On the headless image, strange build error ocurrs:
util.c:(.text+0x14e): undefined reference to `lstat64'
/home/eugen/buildroot2004/buildroot-at91/output/host/bin/arm-buildroot-linux-gnueabihf-ld: /home/eugen/buildroot2004/buildroot-at91/output/build/bluez5_utils-5.54/src/.libs//libshared-mainloop.a(crypto.o): in function `bt_crypto_g2':
crypto.c:(.text+0x74a): undefined reference to `__aeabi_uidivmod'

This is due to bad linkage, bad LDFLAGS, missing -lc and -lgcc
and misconfiguration of the linker.

To avoid this, removed the two steps build and used
a single command with -gcc to build the apps.

Fixes: 0f7bb53 ("package: websocket/wilcbtapps: introduce two new bluetooth/web based apps")
Signed-off-by: Eugen Hristev <eugen.hristev@microchip.com>