sama5: post_build.sh: Create symbolic link when the package is available
Do the libcryptoauth.so symbolic links only when the
BR2_PACKAGE_CRYPTOAUTHLIB is selected. Copy the cryptoauthlib device
configuration only when the BR2_PACKAGE_CRYPTOAUTHLIB is selected.

The update of the /var/lib/cryptoauthlib/0.conf file is not put under
the BR2_PACKAGE_CRYPTOAUTHLIB because the BR2_PACKAGE_CRYPTOAUTHLIB_*
already depends on it. This way we avoid a level of indentation.

While here modify the echo comment to indicate to correct python version.

Signed-off-by: Tudor Ambarus <tudor.ambarus@microchip.com>