sama5: post_build: gg: Mount cgroups in the post build script
Instead of setting Cgroups at every run of the greengrass script,
do it once in the post build script.

We mount the cgroups in the sama5 path and not in the sama5d2, and we
condition the mounting by the presence of the
BR2_PACKAGE_AWS_IOT_GREENGRASS_CORE. This way we can easily extend the
AWS GG support for other sama5 boards if ever needed.

Signed-off-by: Tudor Ambarus <tudor.ambarus@microchip.com>