cryptoauthlib: Bump to v3.2.5
Without configuring cryptoauthlib-v3.2.5 for PKCS11, one sees:
pkcs11_init: PKCS #11 initialization error.
warning: no token URL was provided for this operation; the available tokens are:

https://github.com/MicrochipTech/cryptoauthlib/wiki/PKCS11-Linux-Setup
fails to specify how to configure the library for PKCS11 linux.

The config options were added by comparing the support between what
was on pkcs11 branch and what is available in v3.2.5. This may be
error-prone (not choosing the perfect configuration), but facing the
lack of documentation, choose what seems right. Tested the AWS
greengrass demo on sama5d27_wlsom1_ek, everything was fine.

Signed-off-by: Tudor Ambarus <tudor.ambarus@microchip.com>
Reviewed-by: Nicolas Ferre <nicolas.ferre@microchip.com>