package/uart_brk_ioctl: adapt to 2021.02 requirements
According to buildroot upstream 2021.02 specs:

Download: Rework the tarball creation logic (E.G. when a
        package is fetched from a git/svn repo) to no longer require a
        host-tar <= 1.29. To differ between the old and the new
        format, the new tarballs have gained a -br1 suffix
        (<pkg>-<version>-br1.tar.gz).

The old style of getting a package was adding a '-br1' suffix to the
tarball, which would make the hash check fail for this package.
Update to a simpler github get method, and the checksum works fine,
buildroot does not add -br1 to the tarball.

Signed-off-by: Eugen Hristev <eugen.hristev@microchip.com>