configs: linux.config: Enable HMAC software implementation
The SHA IP implements HMAC acceleration. It's a good idea to enable
at least the same amount of algs as the hardware IPs are supporting.
Provide the posibility for the HMAC hw accelerated crypto algs to
fallback to software implementation in case they need it.
It also helps the crypto test manager to run tests and compare the
results from the software and hardware accelerated implementations.

Signed-off-by: Tudor Ambarus <tudor.ambarus@microchip.com>