package: add hart software service payload generator
The Hart Software Service (HSS) package is not being built.
Buildroot External uses the Hart Software Service's (HSS) payload
generator tool. This tool creates a formatted payload image for
the HSS zero-stage bootloader on PolarFire SoC, given a
configuration file and a set of ELF binaries. The configuration
file is used to map the ELF binaries or binary blobs to the
individual application harts (U54s). Add the HSS payload generator
package to support this.

Signed-off-by: Jamie Gibbons <jamie.gibbons@microchip.com>
Reviewed-by: Conor Dooley <conor.dooley@microchip.com>