configs: icicle: support additional overlay configs
When working on development of adding dt-overlay configurations for
the icicle kit, originally a change to the dt-overlay-mchp.mk had been made
to make these configurations work (see v1/2). Upon further development, it
was discovered that a typo in the dt-overlay Makefile for the dtbo object
variables was the reason this problem.
Upon fixing this error, it highlighted that a different structure of the images
directory with the .dtbo files. Update the post image script to allow the same
functionality as before with this change in folder structure reflected. This
update allows additional configurations to be used correctly without the previous
changes to the dt-overlay-mchp.mk (see v1/2).

Signed-off-by: Jamie Gibbons <jamie.gibbons@microchip.com>