configs: sama5d3-eds: Add NAND Flash build
Adding a default config for producing a NAND flash build for the
SAMA5D3-EDS board.  There is a separate default uboot environment
variables file, post build and post image scripts, as well as NAND
flash QML scripts to use with sam-ba to flash the generated image.
Note that 2 shortcuts are created in the base directory called
flash0 and flash1 to flash the image using ttyACM0 and ttyACM1,
respectively.

Signed-off-by: Jerry Ray <jerry.ray@microchip.com>