board: icicle: genimage.cfg: drop unnecessary properties and entries
Drop the size and offset properties from the U-Boot partition defined in
the genimage configuration file, as there is no need to specify a fixed
size for the payload.bin file. The genimage tool will generate the
partition layout based on the actual size of payload.bin.

Also, drop the u-boot binary from the vfat partition, since the u-boot
binary is included as part of the hss payload binary.

Signed-off-by: Valentina Fernandez <valentina.fernandezalanis@microchip.com>
Reviewed-by: Conor Dooley <conor.dooley@microchip.com>