boards: icicle: add support for generating nor and nand flash images
Add support for generating a nor.img image that can be flashed to a
Micron MT25QL256 NOR flash memory.

Also, add support for generating a nand.img image that can be flashed to
a Winbond W25N01GV NAND flash memory.

For more information on booting from qspi please refer to the
buildroot-external-microchip README and the "Booting from QSPI"
documentation in the polarfire-soc-documentation repository:

- https://github.com/linux4microchip/buildroot-external-microchip
- https://mi-v-ecosystem.github.io/redirects/booting-from-qspi_booting-from-qspi

Signed-off-by: Valentina Fernandez <valentina.fernandezalanis@microchip.com>
Reviewed-by: Conor Dooley <conor.dooley@microchip.com>
Tested-by: Jamie Gibbons <jamie.gibbons@microchip.com>