configs: icicle: add openssl dependancy
Some missing dependancies for PolarFire SoC's icicle kit support were
found during development. The package libssl-dev used to be downloaded
for development in the Buildroot SDK. Instead of using this method, it
was found that we could add this to our defconfigs instead of advising
customers to download the additional dependancy in the buildroot
external readme. Add openssl packge via configs.

Signed-off-by: Jamie Gibbons <jamie.gibbons@microchip.com>
Reviewed-by: Conor Dooley <conor.dooley@microchip.com>