README: add instructions for rootfs and bmaptools for PolarFire SoC
Update the PolarFire SoC section to describe two methods of copying
the image to the eMMC/SD card. One using standard dd and the other using
bmap tools (recommended).

Add the icicle_rootfs_defconfig to the README, which can be used to
generate an image with an ext4 root filesystem.

Signed-off-by: Valentina Fernandez <valentina.fernandezalanis@microchip.com>