mpfs_icicle: add overlays for use at runtime
Buildroot currently only assembles the overlays to the FIT image and not
to the file system. This causes the dt-overlay-mchp example to not
function in Buildroot. Add the overlays to the /boot directory of the
root file system.
The dt-overlay-mchp example uses the /sys directory to modify the device
tree. Mount /sys to avoid permission error when running this script.

Signed-off-by: Jamie Gibbons <jamie.gibbons@microchip.com>
Signed-off-by: Conor Dooley <conor.dooley@microchip.com>