mpfs_icicle: update kernel fragments for v6.1
OF_OVERLAY and OF_CONFIGFS are now enabled in the new mpfs_defconfig for
kernel version 6.1. Remove these configs.
With the update of the Raspberry Pi (RPI) Sense Hat dtso for 6.1, the
sensehat-core module has been removed. The MFD_SIMPLE_MFD_I2C module replaces
the sensehat core one and is selected with the JOYSTICK_SENSEHAT config. Remove
this config.
By default with 6.1, the IIO_ST_MAGN_SPI_3AXIS is configured when the
IIO_ST_MAGN_3AXIS config is selected. This causes warning messages when
using the RPI Sense Hat. Deselect this config.

Signed-off-by: Jamie Gibbons <jamie.gibbons@microchip.com>
Signed-off-by: Conor Dooley <conor.dooley@microchip.com>