configs: enable nfs-utils package
The NFS client is used to mount remote folder in DUT and thus helpful for
testing new kernels and to deploy modules for the matching kernel version.
NFS server features on the other hand are not very useful on boards with
limited resources, which are unlikely to be used as an NFS server.

Signed-off-by: Nayab Sayed <nayabbasha.sayed@microchip.com>
Acked-by: Hari Prasath Gujulan Elango <hari.prasathge@microchip.com>
Signed-off-by: Conor Dooley <conor.dooley@microchip.com>