boards: icicle: add boot script and fitImage to UBI filesystem
Copy the U-Boot boot script and fitImage to the /boot directory in the
UBI filesystem for icicle_nand_defconfig and icicle_nor_defconfig.
This is particularly useful for the QSPI nand and nor images,
since we can now use distro boot to automatically load a boot script
from a UBI filesystem instead of using a custom QSPI boot target.

Signed-off-by: Valentina Fernandez <valentina.fernandezalanis@microchip.com>
Tested-by: Prajna Rajendra Kumar <prajna.rajendrakumar@microchip.com>
Signed-off-by: Conor Dooley <conor.dooley@microchip.com>