configs/icicle: revert systemd support in initramfs icicle defconfigs
The icicle_defconfig and icicle_amp_defconfig include an initramfs
embedded into the kernel. In commit 0202b36, systemd support was added
to all icicle defconfigs. However, having systemd as part of an initramfs
increased the size of the kernel considerably. This is causing the
initramfs not to fit into the memory region we load the kernel+initramfs
image to.

Revert to busybox + mdev instead of systemd for initramfs based
defconfigs, specifically icicle_defconfig and icicle_amp_defconfig.

Please note there is an icicle_rootfs_defconfig which uses a fill ext4
rootfs instead of initramfs and which supports systemd.

Signed-off-by: Valentina Fernandez <valentina.fernandezalanis@microchip.com>