configs: disable auto-loading of WILC SPI module
Remove 'wilc_spi' overlay from 'boot_cmd' U-Boot variable to disable
auto-loading of the WILC SPI driver at boot-up. Whether the WILC SPI board
is connected with the host or not, the driver module gets auto-loaded and
interferes with SDIO testing. With this change, the 'wilc_spi' overlay is
not loaded in the default image but it can be enabled at run-time for
testing the SPI mode.

Signed-off-by: Ajay Singh <ajay.kathat@microchip.com>
Reviewed-by: Nayab Sayed <nayabbasha.sayed@microchip.com>
Signed-off-by: Jamie Gibbons <jamie.gibbons@microchip.com>