package: mpfs_amp_examples: bump package for release prep
Update MPFS AMP examples with the latest hash for 2024.06 release
preparation. This includes updates to the mpfs-rpmsg-bm and
mpfs-rpmsg-freertos examples, which now use the Mi-V IHC IP v2.

The Mi-V IHC IP version 2 is not backwards compatible with the Mi-V IHC
subsystem used in the Icicle Kit reference design 2024.02 or earlier.
For this reason the AMP support in the v2024.06 release will not be
backwards compatible with previous releases.

Signed-off-by: Jamie Gibbons <jamie.gibbons@microchip.com>