icicle: make rootfs config default
The "default" configuration used for Icicle kits in Buildroot is the
initramfs configuration. However, the most commonly used configuration
is the rootfs configuration. It has been decided that this should be the
default configuration moving forward as it gives you the most feature
rich and least resistricted option. Make the rootfs configuration the
"default" Icicle defconfig.

Signed-off-by: Jamie Gibbons <jamie.gibbons@microchip.com>
Tested-by: Valentina Fernandez <valentina.fernandezalanis@microchip.com>
Signed-off-by: Jamie Gibbons <jamie.gibbons@microchip.com>