configs/icicle: add configs to support saveenv cmd in qspi defconfigs
To support the U-Boot `saveenv` command in QSPI configurations a custom
ubinize configuration file is used to set up two UBI volumes, the first
of which contains the U-boot environment variables.
To use the U-Boot `saveenv` command without generating a warning at the
inital boot we need to generate an existing env binary in the fat
partition. Use the U-Boot tool `mkenvimage` to do this.

Acked-by: Valentina Fernandez <valentina.fernandezalanis@microchip.com>
Signed-off-by: Jamie Gibbons <jamie.gibbons@microchip.com>