configs: sama7g5ek_optee: update to use patches for the devicetree
Remove the path of dt files from "BR2_LINUX_KERNEL_CUSTOM_DTS_PATH" due to
the dt files was removed. The dt files in the source code of the kernel
would be used then.
As the patches for the devicetree are added, update "BR2_GLOBAL_PATCH_DIR"
by adding the path of the patches.

Signed-off-by: Tony Han <tony.han@microchip.com>
Reviewed-by: Nayab Sayed <nayabbasha.sayed@microchip.com>
Tested-by: Charan Pedumuru <charan.pedumuru@microchip.com>
Signed-off-by: Jamie Gibbons <jamie.gibbons@microchip.com>