package/egt*: bump versions for linux4microchip-2025.04
Bump the following packages for the 2025.04 release of linux4microchip:

egt to 1.11-rc3
egt-benchmark to 1.3-rc1
egt-launcher to 1.5.1
egt-samples to 1.7-rc1
egt-samples-contribution to 1.3-rc1
egt-thermostat to 1.8.1
libplanes to v1.1.0-rc1

Update the egt package as this new version offers soundeffect support
that requires alsa and sndfile libraries. Update the minimal requirement
for gcc as egt no longer uses c++14, but c++17.

Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>
Reviewed-by: Nayab Sayed <nayabbasha.sayed@microchip.com>