package/libm2d: reintroduce libm2d
libm2d has been reworked and provides an API for 2D acceleration that is
used by EGT. It supports the GFX2D accelerator found on SAM9X60 and
SAM9X75.

Remove the vestige of the dependency of cairo on libm2d. It is no longer
true with this new version.

Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>
Reviewed-by: Nayab Sayed <nayabbasha.sayed@microchip.com>
Signed-off-by: Conor Dooley <conor.dooley@microchip.com>