board: microchip: Recalculate and re-sign IPA modules
Buildroot strips its binaries post-installation as an optimization,
which invalidates the original IPA module signatures done during
installation. This causes libcamera to treat them as closed-source and
attempt to run them in isolation using IPC, which fails if the modules
aren't designed for isolated execution/doesn't have IPC
support.Recalculating and re-signing the IPA module signatures is
therefore necessary to ensure proper functionality.

Add support for this in 'post-build.sh', conditionally if the libcamera
IPA module is selected. This ensures IPA modules are signed correctly
after they are built and installed.

Signed-off-by: Balakrishnan Sambath <balakrishnan.s@microchip.com>
Signed-off-by: Conor Dooley <conor.dooley@microchip.com>