package/mpfs_amp_examples: add support for Mi-V IHC IP v2
Bump the mpfs_amp_examples makefile to include updates to the
mpfs-rpmsg-bm and mpfs-rpmsg-freertos examples, which now use the
Mi-V IHC IP v2.

The Mi-V IHC IP version two is not backwards compatible with the
Mi-V IHC subsystem used in the Icicle Kit reference design 2025.03
or earlier. For this reason the AMP support in the v2025.07
release will not be backwards compatible with previous releases.

Signed-off-by: Jamie Gibbons <jamie.gibbons@microchip.com>
Reviewed-by: Valentina Fernandez <valentina.fernandezalanis@microchip.com>