package/dt-overlay-mchp: fix install target cmd macro
The DT_OVERLAY_MCHP_INSTALL_TARGET_CMDS macro foreach loop uses
wildcard to find all .dtbo files in any subdirectory of $(@D). If the
.dtbo file was found in a different subdirectory (not
$(BR2_PACKAGE_DT_OVERLAY_MCHP_PLATFORM)), the install command will look
for it in the wrong place and fail. Fix this to ensure it only installs
files from the correct subdirectory. This also allows for the use case
where there is no dtbo subdirectory for a given .its.

Signed-off-by: Jamie Gibbons <jamie.gibbons@microchip.com>
Acked-by: Conor Dooley <conor.dooley@microchip.com>