package/dt-overlay-mchp: fix install dtbo to target loop
Buildroot parses the package .mk file before the package source is
downloaded. Due to this, the .dtso nor .dtbo files exist before the
wildcard assignment is evaluated. This means wildcard assignments for
files in the souce directory do not work as expected and cause .dtbos to
not be installed when they do in fact exist.
Therefore, using a shell loop is the Buildroot recommended and most
robust way to ensure .dtbos - if they exist - be installed.

Signed-off-by: Jamie Gibbons <jamie.gibbons@microchip.com>
Acked-by: Conor Dooley <conor.dooley@microchip.com>