board: microchip: icicle: add OF_BOARD config
The HSS puts a minimal dtb in it's payload's ancillary-data immediately
after U-Boot in memory. CONFIG_OF_BOARD will use this dtb that the HSS
carries with the bare minimum of nodes enabled. It allows for firmware
to provide the address of the devicetree in memory using the `a1`
register.

Signed-off-by: Jamie Gibbons <jamie.gibbons@microchip.com>
Reviewed-by: Valentina Fernandez <valentina.fernandezalanis@microchip.com>