package/mpfs_examples: and kconfig option to set example directories
The MPFS example directories are currently set by the package .mk file.
This means that once the package is enabled the same examples get
installed for all Buildroot board defconfigs. Add a kconfig option to
set the examples installed from the Buildroot board defconfig. This
allows different examples to be installed for different boards and
configurations, as not all examples are supported by all boards.

Signed-off-by: Jamie Gibbons <jamie.gibbons@microchip.com>
Acked-by: Valentina Fernandez <valentina.fernandezalanis@microchip.com>