package/pic64gx_amp_examples: add package
Add PIC64GX AMP examples, the current example (remote firmware),
compiled by default demonstrates the usage of the OpenAMP API and the
IHC driver to communicate between Linux and the remote firmware - the
example.
Please note that this package generates a remote firmware that can only be loaded
after Linux boot, as opposed to being loaded by the HSS. Another build
with a different configuration is necessary for that functionality.

Signed-off-by: Pierre-Henry Moussay <pierre-henry.moussay@microchip.com>
Signed-off-by: Jamie Gibbons <jamie.gibbons@microchip.com>