pic64gx_curiosity_kit_amp: add missing systemd growfs functionality
Currently, the disk usage is at 100% for the PIC64GX Curiosity kit AMP
image. Implement growfs to grow the size of the root partition on the
first boot. This allows for a smaller image size. Repart uses the fstab
and repart.d configurations to repartition the disk and subsequently
grow the root filesystem using growfs. Add the systemd growfs and repart
config files to support this.
The partition UUIDs are incorrect and thus causing growfs to incorrectly
attempt to resize the partitions. Update the UUIDs to the correct
partition types to fix this and allow growfs to work as expected.

Signed-off-by: Jamie Gibbons <jamie.gibbons@microchip.com>
Tested-by: Valentina Fernandez <valentina.fernandezalanis@microchip.com>
Signed-off-by: Jamie Gibbons <jamie.gibbons@microchip.com>